refactor(#164): annotations/holes.py — hole/pattern annotation pass (P5d)
Move the largest annotation capability into annotations/holes.py: per-hole
callouts and balloons (_annotate_holes), location dimensions including side-drilled
holes (_add_location_dims/_locate_off_axis_holes, #133), pitch/grid pattern dims
(_add_grid_pitch_dims/_place_pitch_dim/_add_furniture), slots (_annotate_slots),
plus the holes-only helpers _legible_locations/_record_callout_drop/
_solve_strip_via_layout.

holes.py imports the shared placement helpers from annotations._common, the strip
solver from draftwright.layout, and _core/build123d/features — never annotate/
make_drawing, so the orchestrator calls down with no cycle. The orchestrator-only
helpers (_wrap_rows, _is_concentric_hole, _concentric_bore_diams,
_detect_step_repeat, _maybe_tabulate_holes) stay with _auto_annotate. annotate.py
re-imports the four passes _auto_annotate calls; two tests now import the moved
helpers from holes.py.

Pure move: golden gate passes against the committed snapshots with no regeneration;
full make_drawing/golden suites pass (357); ruff check + format + mypy clean.
Adversarially checked: no upward import, no stale defs, holed/patterned parts get
callouts + centre marks + location dims, lint clean.

Part of #164 (P5e orchestrator + envelope closes it).
